Contest service

+++ leaderboard/answer.go
@@ -2,6 +2,6 @@ package leaderboard
 
 type Answer struct {
     Id      int
-    Created int
+    Created int64
     UserId  int
 }
